simple-examples/pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.0"
description = "Easily run background tasks in Python with Hatchet"
authors = []
readme = "README.md"

[tool.poetry.scripts]
hatchet = "src.main:start"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
hatchet-sdk = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

